refactor: refactor profile using new profile-layout components
