+++ src/components/profile/profile.module.css
@@ -1,13 +1,3 @@
-.container {
-  position: relative;
-}
-
-.sidebar {
-  max-width: 320px;
-  position: absolute;
-  left: -380px;
-}
-
 .input-field input {
   color: #8585ad;
 }
